Extract module for value storing/loading
